添加公共场地接口完成

+++ src/main/java/com/backend/tjtablepartyspringboot/dto/PublicSiteDto.java
@@ -33,6 +33,6 @@ public class PublicSiteDto {
     Date uploadTime;
     Date checkTime;
     String[] type;
-    byte status;
+    int status;
     List<PublicSiteTimeDto> openTime;
 }
